floatval

+++ includes/mb/class.lmsensors.inc.php
@@ -87,7 +87,7 @@ class LMSensors extends Sensors
             if (count($data)>2) {
                 foreach ($data as $key=>$value) {
                     if (preg_match("/^\+?(-?[0-9\.]+).?$/", trim($value), $newvalue)) {
-                        $data[$key] = intval($newvalue[1]);
+                        $data[$key] = floatval($newvalue[1]);
                     } else {
                         $data[$key] = trim($value);
                     }
@@ -190,7 +190,7 @@ class LMSensors extends Sensors
             if (count($data)>2) {
                  foreach ($data as $key=>$value) {
                     if (preg_match("/^\+?(-?[0-9\.]+).?$/", trim($value), $newvalue)) {
-                        $data[$key] = intval($newvalue[1]);
+                        $data[$key] = floatval($newvalue[1]);
                     } else {
                         $data[$key] = trim($value);
                     }
@@ -244,7 +244,7 @@ class LMSensors extends Sensors
             if (count($data)>2) {
                 foreach ($data as $key=>$value) {
                     if (preg_match("/^\+?(-?[0-9\.]+)$/", trim($value), $newvalue)) {
-                        $data[$key] = intval($newvalue[1]);
+                        $data[$key] = floatval($newvalue[1]);
                     } else {
                         $data[$key] = trim($value);
                     }
@@ -307,7 +307,7 @@ class LMSensors extends Sensors
             if (count($data)>2) {
                 foreach ($data as $key=>$value) {
                     if (preg_match("/^\+?(-?[0-9\.]+).?$/", trim($value), $newvalue)) {
-                        $data[$key] = intval($newvalue[1]);
+                        $data[$key] = floatval($newvalue[1]);
                     } else {
                         $data[$key] = trim($value);
                     }
